Changed the HTTP tests to be sectioned out into unit and integration
